requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
idna==3.2
pymongo==3.12.0
pytz==2021.1
requests==2.26.0
sqlparse==0.2.4
urllib3==1.26.6
